make filterablelistmixin respect Wagtail sites
This change modifies how the `FilterableListMixin` page mixin class
selects and limits the pages that it filters, making sure that it only
filters pages that are in the same Wagtail Site as itself.

Currently this mixin filters any live `AbstractFilterPage`, even those
that are elsewhere in the page tree. This causes problems if, for
example, a live page is moved to the Trash, or to some other Site
running in the same Wagtail instance.

With this change, the default behavior of Page classes that inherit from
`FilterableListMixin` is to only include pages in their same Site.

This change includes some minor refactoring of how the filterable
queryset gets passed around, renaming some variables `filterable_` for
clarity and using the page's own `Page.get_site` method for
convenience. Some logic around page category filtering was also
collected in one place, with new tests added.

The default `FilterableListMixin` behavior continues to filter only
direct children of the page, and use 10 results per page. This behavior
can now be overridden for inherited classes (like `NewsroomLandingPage`)
using some class attributes; this removes some code duplication.

One side effect of this change is that any filterable page that lives
outside of any Site will be unable to filter any pages. This seemed like
a reasonable decision to me: if a page isn't part of a Site, then it's
unclear where it should get its filterable pages from.