Always sort lists to keep ordering stable, avoid potential test breakage
When running the tests in a Docker container (the filesystem might
be behaving differently, which is my only guess as to why this has
not come up before), I noticed that this test would break sometimes
because of the ordering of the items in the first list. When I
sort the lists, I can no longer observe the problem.