bump pyelasticsearch requirement to 0.6.1
This commit bumps the pyelasticsearch requirement from 0.6 to 0.6.1.

This package is only used by the optional regulations-core app:

https://github.com/search?q=user%3Acfpb+pyelasticsearch&type=Code&utf8=%E2%9C%93

Ideally this would only be defined in that repository's setup.py, but
currently (unfortunately) we include all satellite app requirements in
this repository's requirements files.

The use of pyelasticsearch==0.6 was causing an issue because that
version enforces requests<2.0,>=1.0, which conflicts with the
requests==2.12.4 that we have in our requirements.

The relevant difference between pyes 0.6 and 0.6.1 is adding support for
requests<3.0. See changelog at:

https://github.com/pyelasticsearch/pyelasticsearch/blob/01b0cb7671e81556c28ddb0d2c74393b5a63ad80/docs/source/changelog.rst

Thanks to @acrewdson for discovering this incompatibility.