Use venv-update (w/pip-faster) to speed up tox
Using venv-update/pip-faster, we are able to keep the tox
virtualenv and not destroy-and-create before each run. This
is because, unlike tox's default behavior, with venv-update
we get automatic syncing of `pip -r` deps: new deps are
installed and unused ones are removed. This way we get much
faster `tox` times, not having to wait the several minutes
it can take to install all the deps in a fresh virtualenv.

NB: tox is being upgraded here because the newer version is
required by tox-pip-extensions.