Downgrade moto to address Jinja2 version conflict
Context here is that we're wanting to get rid of all dependency
version conflicts so that we can use pip replacements that are
stricter than stock pip in this respect (pip-faster or pip-compile,
neither of which tolerate dep version conflicts).

Going back to moto 0.4.27 allows us to remain on the version of
Jinja2 specified in requirements/base.txt. If/when this PR

    https://github.com/spulec/moto/pull/1474

gets merged, then we can consider bumping moto to the latest
version (a version containing the change in that PR) and see
if it works with our tests. Until then, we'd need to stay at
version 0.4.27.