deprecate python unit testing of optional apps
In the past we tried to enable running the Python unit tests of our
optional apps (from requirements/optional-public.txt) using the settings
of this cfgov-refresh project.

This practice has fallen out of favor, and instead we now try to make
sure that optional packages themselves have robust and comprehensive
tests, which negate the need for testing within cfgov-refresh.

This change removes the tox `optional` and `optional-fast`  environments
that supported this kind of testing.

This change also removes the `OptionalAppsMixin` mixin class that we have
in our test runner that enables testing of those apps. Motivation for
this change is to pursue simplification and removal of unused code.