better testing and development with govdelivery
This commit refactors how the GovDelivery Python client is interfaced
with from code and adds some helpful testing classes for local
development and unit testing.

The cf.gov website uses the GovDelivery platform
(https://insights.govdelivery.com/Communications) to subscribe users to
mailing lists. We use the `govdelivery` Python package (our own,
published to PyPI with source code at
https://github.com/cfpb/govdelivery) as a client for the GovDelivery
API.

Interfacing with this API requires a username and password, and
subscribing users to mailing lists requires that those lists be created
and have a unique account code. This can make it difficult to test
mailing list signup both when running the site locally during
development as well as for unit tests.

The current implementation requires developers to have a valid set of
credentials and have mailing lists on a GovDelivery staging site in
order to fully test locally; it also requires configuration of various
environment variables to support this. Unit testing is done with mocks
and patches against the `govdelivery` code.

This commit instead provides alternate Python classes that can be used
in place of the real `govdelivery.api.GovDelivery` class. In
development, a replacement class can be used to instead log API requests
to the console, making it much easier to test signup flow. For unit
testing, several classes are provided that simulate valid and invalid
responses from the API.

The new functionality is created in `core.govdelivery`, and documented
in the site documentation under "Backend testing". All new code has full
unit test coverage. Existing GovDelivery interfaces (exactly two: in
`core.views` and in `data_research.handlers` have been updated to use
this changed approach, as have their unit tests.

These changes were motivated by some ongoing work on the conference
registration form that interfaces with GovDelivery, and a desire to
improve the ability to develop and test against that form.

To see this change running locally, run a local version of the site and
visit a page with an email signup form, for example
http://localhost:8000/owning-a-home/. You'll see that if you submit the
form you get console output indicating how the GovDelivery API was
called.

Note that default behavior in production should stay the same with this
change as the API is still used in the same way.