delete and recreate django migrations
This commit deletes and recreates all Django migrations for the various
apps in this project. It generates a similar outcome to how built-in
Django squashing works, but works better with the large backlog of
RunPython commands we've used over time.

This should significantly speed up the time it takes to create a new
test database (i.e. using `tox`), although, unfortunately, this is still
somewhat slow due to the number of models and migrations in third-party
apps like Wagtail.

It has the effect of recreating a set of new migrations that describe
the current state of all of our models, with these critical side
effects:

1. Databases that exist at some migration state before the one at the
point of this commit (7fddfdd) will no longer be able to be migrated
to this current state, as the intermediate changes will have been lost.
This means that those databases will need to be recreated.

2. Any open pull requests that reference or depend on some of the
existing migrations will need to be modified to instead refer to these
new migration files.

For the record, these new migrations were generated using this approach:

1. Remove all existing migration files:

```sh
$ rm -f -v cfgov/*/migrations/0*
```

2. Create new migration files from the state of model Python code:

```sh
$ cfgov/manage.py makemigrations --noinput
```

As it happens this creates new initial migrations (`0001_initial`) for
all apps, plus some subsequent migrations (`0002_something`) for apps
that depend on other apps (for example, `ask_cfpb` has its own initial
migration and then some changes that rely on `v1`).

3. Rename the created migration files so that they follow in sequence
the migration files that used to exist. For example, currently there are
101 `v1` migrations, so the first new migration is numbered 102.

4. Manually alter all new migration files to indicate that they replace
the old migration files.

This involves adding lines to these files like:

```py
includes = [('app_name', '0002_foo'), ('app_name', '0003_bar'), ...]
```

This tells Django that these new files replace the old files, so that
when migrations are run again, it doesn't need to do anything.

Also manually update any new subsequent migration files so that they
properly refer to each other. For example, if an app has two new
migrations 102 and 103, the 103 file needs to properly depend on 102.

To apply these new migration files to an existing database, you can
simply run:

```sh
$ cfgov/manage.py migrate --noinput
```

You'll see that there are no changes to apply, as the new files should
exactly describe the current model state in the same way that the old
migrations did.