fix related posts urls for duplicate slugs
The current logic in the Related Posts template is using the built-in
Wagtail
[`slugurl`](http://docs.wagtail.io/en/v2.0/topics/writing_templates.html#slugurl)
template tag, which returns an undetermined link if there are multiple
pages with the same slug anywhere in the site.

This change fixes GHE#2533 which was generating an incorrect link due to
two pages having the same slug. It instead uses the built-in `page.url`
property which returns the best URL for the page.

Note that due to the current way we have our production Wagtail Sites
configured (we have multiple), `page.url` may return a full absolute URL
rather than a relative URL here. I think the downside of this is
outweighed by the simplicity of the required code change. Otherwise we'd
have to put in some complex logic like
`page.relative_url(page.get_site())` which I would rather not stick in
the template.

If, hopefully, in future we move to a single Wagtail Site in production
these links will properly be relative ones.

To test, use a production dump and visit
http://localhost:8000/data-research/ and observe that the "2018 CFPB
Research Conference" link properly points to an event page, not the
conference landing page.