apply download icon to links regardless of case
The current logic that applies the download icon
(class="cf-icon-download") to certain links only works if the link file
extension is all lowercase. For example, "foo.pdf" will get the icon but
"foo.PDF" will not.

The list of extensions that triggers the download icon is defined in
`settings.DOWNLOAD_LINK_PATTERN`. This change adds case insensitivity
`(?i)` to the regular expression that matches against links, so that all
files with the specified extensions get the icon.

See the new unit test that validates this behavior. Additionally, using
a production dump you can visit
http://localhost:8000/about-us/blog/what-do-if-youre-wrongfully-billed-medicare-costs/
and notice that the link "printer-friendly-version" towards the bottom
of the article properly gets the download icon, even though it links to
a file ending in `.PDF`. This is better than the current production
behavior at
https://www.consumerfinance.gov/about-us/blog/what-do-if-youre-wrongfully-billed-medicare-costs/
which is missing the icon.

This change fixes platform#788.