support DATABASE_URL to customize test db
The Python unit tests in this project use a SQLite database. It would be
useful to support alternate databases, for example for testing test
compatibility with Postgres.

This change allows users to define the `DATABASE_URL` environment
variable to test against different databases, for example:

```sh
$ DATABASE_URL=postgres://user@localhost/cfgov tox -e fast
```

This makes use of the
[dj-database-url](https://github.com/kennethreitz/dj-database-url)
library.

I've updated the manual to describe this change.

Note that in future it might be interesting or desirable to consider
simplifying our DB configuration so that we always use a database URL.
For example, tox could pass in different URLs for testing instead of
having numerous settings files with different DBs. I considered this but
decided to keep the scope of this PR small.