Ask CFPB: Exclude redirected pages from index
Redirected pages need to left in the "live" state so that they can
redirect themselves if an editor sets up a redirect. But their live status
left their related question open to being indexed, which allowed
the redirected answer to show up in search results.

This adjusts the query used to create the elasticsearch index so that
redirected pages are excluded.

Since we index answer objects, and we need to consult the related pages
to determine whether they are redirected, the query is relatively
expensive. But that's not an issue here, because the query only runs
once at index time, and that is once a day.

I confirmed locally that answer 1535, which is currently the first
result for the query "spouse credit card debt," disappears from search
after re-indexing with this new indexing query, along with 147 other
redirected answers.

The index query for Spanish answers is updated as well.