Remove tests that rely on how NULLs are sorted
These tests promise functionality that is specific to
MySQL.  With our impending move to PostgreSQL, the behavior
in these two tests will no longer be the case (sorting NULLs first).
However, we don't actually rely on this ordering anywhere, so we can
just remove the tests and not promise this behavior.