cleanup references to TOXENV env variable
As a followup to #3894, the use of the `py27` environment for tox is now
deprecated and so references to it should be removed. The invocation of
just plain `tox` now aliases to both `lint-py27` and the equivalent
`unittest-py27-dj18-wag110-slow` environment.

TODOS: remove `TOXENV` if you have it in your `.env` file, unless you
want to take the opportunity to set it to `fast`.