fix jobmanager unit tests db usage
This commit modifies some of the jobmanager app's unit tests so that
they properly rollback the DB changes that are made during tests.
Instead of inheriting from `unittest.TestCase`, these tests need to
inherit from `django.test.TestCase` so that the database changes happen
in a transaction.

This fixes the usage of these tests when running against a Postgres DB.
This can be verified by running, e.g.:

```sh
$  DATABASE_URL=postgres://user@localhost/cfgov tox -e fast jobmanager.tests
```

(This requires the fix to the Tox config in #3954.)

This PR also fixes a tiny bug (which may not have been causing the
failures) around not properly calling the parent `setUpClass`.