fail frontend.sh script if not using node 8
This change modifies the frontend.sh script so that instead of only
warning if a version of Node besides 8 is being used, it fails.

This behavior was originally introduced as a failure in #3177 but then
removed in #3178 due to Jenkins version incompatibility. Jenkins is now
properly using Node 8 so this can be changed back to a failure.

If a user isn't using Node 8, there's no point in continuing the script,
as the build will fail.