remove deprecated simplejson requirement
This change removes a deprecated dependency on
[simplejson==3.8.2](https://pypi.python.org/pypi/simplejson/3.8.2).

There's one remaining reference to this library in the optional
[cfpb/retirement](https://github.com/cfpb/retirement) app; however that
reference is only used as a fallback when `json` is not defined. That
was added in Python 2.6 and this repository requires Python 2.7, so that
fallback is no longer needed. Additionally that code only exists as part
of some browser testing code:

https://github.com/cfpb/retirement/blob/46f0610392fd94fcf52e87cf83275c63f75e55c7/browser_testing/features/environment.py#L17

This library was used in various other libraries which are no longer
part of cfgov-refresh ([see search
here](https://pypi.python.org/pypi/simplejson/3.8.2)).