simpler, more flexible database configuration
This set of changes attempts to make this project's Django database
configuration less MySQL-specific, making it easier to run against other
databases like SQLite and PostgreSQL.

Instead of specifying the Django database through a series of
environment variables like `MYSQL_HOST` and `MYSQL_USER`, we want to
standardize on the use of a single `DATABASE_URL` variable. This
variable is supported and documented as part of the third-party
dj-database-url package:

https://github.com/kennethreitz/dj-database-url

This lets you swap between databases by modifying a single variable
from, say, `mysql://user:pw@host/name` to
`postgres://user:pw@otherhost/othername`. This makes it significantly
easier to run this project against different databases and reduces the
number of variables needed in our deployment environments.

The various Django settings files have been condensed based on the
introduction of this variable. The new logic works like this:

1. If `DATABASE_URL` is defined, Django will use it to define the
default database.
2. If not, and if `MYSQL_HOST` is defined, Django will use a MySQL
database as the default. This is to maintain compatibility with the
current behavior. A warning is displayed if this logic is used, however,
because, as part of the site transition to Postgres, support for these
specific variables will be removed in future.
3. If neither of these are set, then there is no fallback database, and
most Django actions will fail with an error. The exception to this is
for unit tests, as described below.
4. When running the unit tests using `settings.test`,
`settings.test_nomigrations`, and `settings.test_acceptance`, special
logic is used. If a default database is already defined as described
above, it'll be used for the tests -- this is different from the current
logic but ensures that e.g. if you're running against Postgres that the
tests will run against Postgres.
5. If no default database is defined via environment variables, for
example upon initial checkout of this repository, an in-memory SQLite
database is used for testing. @sebworks points out that using an
in-memory database is faster than using one on disk.

The requirements files have been modified slightly to treat MySQL and
PostgreSQL more equally. There's now a separate
`requirements/postgres.txt` file with a single `psycopg2` requirement.

The definition of a second Postgres database and a Django router to
support it (introduced in #3462 / 7ebe9a8) has been removed, as the
Postgres conversion project will not be utilizing side-by-side
databases.

Both the Docker (`.python_env`) and non-Docker (`.env`) environment
variable templates have been modified to reflect the preference for
using `DATABASE_URL`. The Docker path in `setup.sh` has been modified to
ensure that `.python_env_SAMPLE` is used to create an initial
`.python_env` if it doesn't exist. The default Docker Compose setup
doesn't contain an explicit database selection, deferring instead to the
environment (as the non-Docker setup does).

I've made a few other minor tweaks to the setup scripts and
documentation to remove a few things that were outdated. I changed the
default database selection in `.env_SAMPLE` from MySQL to SQLite to
simplify development for new users and reduce the startup process. See
the modified quickstart guide in `README.md`.

This PR also modifies `refresh-data.sh` so that it can work on gzipped
archives without expanding them. This reduces the disk space needed when
working with database dumps.

I apologize for the length of this PR; these are all related changes and
I thought it best to introduce them as a set. I'm happy to break out
into smaller PRs if preferred. Note that because the legacy MySQL
environment variables are still supported this will not (yet) require
any changes to the deployment pipeline. Future work will be needed to
standardize on the use of `DATABASE_URL` there.