add new svg_icon template tag to render SVG icons
This change adds a new `svg_icon` template tag for both Django and Jinja
templates, that allows users to render the contents of a static SVG file
in their templates.

Usage in Django templates is like this:

```html
{% load svg_icon %}
{% svg_icon "test" %}
```

And in Jinja templates like this:

```html
{{ svg_icon("test") }}
```

(This requires the filter to be added to the environment in the
`settings.TEMPLATES` Jinja engine.)

In both cases the tag will look for a file named `icons/test.svg` in the
regular Django staticfiles search path.

This PR is intended to implement some groundwork to enable the use of
Capital Framework icons as in #3887.