add ability to mock static files in tests
This change introduces a new helper class for use in Python unit tests:
`core.testutils.mock_staticfiles.MockStaticfilesFinder`. This finder can
be added to the list of Django staticfiles finders here:

https://docs.djangoproject.com/en/2.0/ref/settings/#std:setting-STATICFILES_FINDERS

This class gives you the ability to mock the existence of static files
for test purposes even if those files to not exist in the repository at
test time. Generally this is useful when some other process (e.g. the
frontend build) will be generating those files, and we don't want to
make the backend tests dependent on that process. In particular this PR
is being introduced to support the work in #3887 to add Capital
Framework icons to this repository.

See the documentation and tests for a detailed explanation of how this
code works. As an overview, in test settings you can now add:

```py
from base_settings import *

STATICFILES_FINDERS += [
   'core.testutils.mock_staticfiles.MockStaticfilesFinder',
]

MOCK_STATICFILES_PATTERNS={
    'files/to/mock.jpg': 'files/to/serve/instead.jpg',
}
```

The mock finder acts like the files being mocked exist but return the
path to another file instead; this other file must be findable by other
staticfiles finders you have defined (e.g. Django's `FileSystemFinder`).

Note that unfortunately the Django docs currently say that "Static file
finders are currently considered a private interface, and this interface
is thus undocumented."

https://docs.djangoproject.com/en/2.0/ref/settings/#staticfiles-finders

I've tried to add sufficient tests to verify the intended behavior and
ensure that it continues to work in future Django upgrades.