only warn about deprecated db variables locally
The changes in #4013 (c94b26e) introduced a warning message about
database environment variables being deprecated. Placing this message in
`settings.base` unfortunately means that it is logged even in
production, where we still use those legacy variables.

We still want this warning, but only locally and in tests.

This change moves the warning from `settings.base` to `settings.local`.