Add Regulations 3000
Regulations 3000 is a Django/Wagtail app for storing and presenting the CFPB's regulations. It is intended to replace eRegulations.

Regulations 3000 includes Django models for regulation text. Parts are the part of the Code of Federal Regulations that contains the regulation text, EffectiveVersions the versions of the regulation text that are effective as of specific dates, Subparts are logical groupings of regulation sections, and Sections store the text of a section of the regulation.

Regulations 3000 includes the Wagtail pages RegulationLandingPage, the entry point into Regulations 3000, and RegulationPage, a routable page type that displays a regulation stored in the Django models.

Regulation text is stored as Regdown, an extension of Markdown that includes support for labeled paragraphs, for linking directly to paragraphs, and block references, for allowing interpretations to appear inline with the regulation text being interpreted.