Bump moto to more recent version and make related fixes
Back in 80c7c7e88b4b82953728aedc7156307ea980dd7e we downgraded moto to
0.4.27 to fix some conflicting depdendency versions that had surfaced
through the addition of 'venv-update' to the tox setup. At the time, I
submitted a PR to moto to address the problem with their dependencies,
and that eventually got merged. But we had not yet gotten round to
bumping moto back up to a more recent version.

Now we have no choice: our old version of moto specifies an unpinned
version of flask, and flask released version 1.0 today. In flask 1.0,
the Jinja2 dep is bumped, meaning that we now have another Jinja2
version conflict when attempting to run tox.

All of which is to say, we are now bumping moto back up to a more recent
version, and fixing a few related tests that needed to be updated as a
side-effect. While moto 1.3.1 _is_ a lot more recent (one month old),
it's not the newest version as of this writing (1.3.3). Upgrading to
1.3.3 would mean upgrading boto3, and we might prefer to kick the can
down the road on that one for now.

Also note that we are bumping requests to the latest version to
address a version conflict (newer moto requires newer requests).
Upgrading requests is probably a good idea in general, and should not
pose problems since requests claims minor versions do not introduce
breaking changes.