add mysql-specific refresh script
After the changes in #4013 (c94b26e), the database refresh script will
fail if the Django database user doesn't have sufficient privileges to
import the database dump. For MySQL this can happen if the dump contains
the CREATE ALGORITHM statements that are used for a view. In practice
this shouldn't happen, but as an artifact of the current production
database, this does.

This change adds a new refresh-mysql.sh script that wraps the existing
refresh-data.sh script with a new preface that first ensures that the
Django database user has super privileges.

Once the site transitions to Postgres, this will no longer be an issue.
Either at that time, or when those views are removed from production
data dumps, this script can be removed.

See platform#2711 and comments by @higs4281 there.