adds youth financial education page to beta drafts
This change adds Wagtail page ID 11432 to the set of pages that are
publicly visible in draft on beta. This change makes it so that the
latest draft version of Wagtail page 11432 will be visible at its URL on
beta, even if the page isn't/is never published.

This is to support work to preview the Teachers Digital Platform.