Fix an issue where richtext images are being distorted
Currently richtext images are printed to the markup with static height
and width attributes. When the image is resized in one dirction, it
can be distorted in the opposite direction. Resetting the height and
width to auto avoids this issue allowing the image to take up the
dimensions of it's container without causing distortion.

Updated comment based on feedback