Ensure only reg3k links use turbolinks
The turbolinks JS library is only loaded on reg3k pages. This means
that only reg3k pages are turbolink-ified, right? Not quite. Once you
visit a reg3k page all future page navigations are turbolink-ified because
turbolinks intercepts link clicks and AJAX's new content. The only way
to break out of this behavior is for the user to hard refresh a non-reg3k
page.

The solution is to disable turbolinks by default and enable it for
individual links by explicitly giving them a data-turbolinks attribute.
This ensures that only links we specifically call out will get the
turbolinks AJAX treatment.

https://github.com/turbolinks/turbolinks/issues/231