use postgres version 10 and cfpb username
This change bumps the version of the PostgreSQL database created in
Docker from 9.6 to 10.

It bumps Travis to version 10 as well, using the travis database and
user that are created as part of v10 setup there (see
GH/travis-ci/travis-ci/issues/8537).

It also standardizes the default Postgres username to cfpb, to be more
consistent with the data dumps generated by
https://github.com/cfpb/cfgov-mysql-to-postgres.