use myisam for local mysql and migrate on refresh
This commit fixes two current issues around using a production MySQL
dump in a local MySQL setup.

By way of context, our production MySQL has a default storage engine of
MyISAM. This causes problems if your local MySQL has a default storage
engine of InnoDB and create a new table locally that has a foreign key
to an old table in the MyISAM storage engine. MyISAM doesn't support
foreign keys and so a foreign key like this can cause problems:

https://github.com/cfpb/cfgov-refresh/blob/master/cfgov/regulations3k/migrations/0001_initial.py#L48

In this case, the foreign key from a new table created locally when you
run this migration (which could be in your local default InnoDB) can't
be made to an old table brought in from the production dump which was in
MyISAM.

As a fix, we should always run our local MySQL using MyISAM as the
default storage engine. This will mean that any new tables created via
migrations will also be in MyISAM, which will be consistent with how
things are stored in production.

Note that the Docker version of this change requires setting the
variable storage_engine, because we are using MySQL 5.1 in Docker:

http://download.nust.na/pub6/mysql/doc/refman/5.1/en/storage-engine-setting.html

But the non-Docker version uses the variable default_storage_engine,
because most developers are using MySQL 5.7 locally:

https://dev.mysql.com/doc/refman/5.7/en/storage-engine-setting.html

Somewhat relatedly, there's an issue in the refresh data script because
we don't currently run migrations before we run the initial_data script.
This can cause problems if a new field exists in code after the time
that the production dump was generated. When the initial_data script is
run, it tries to pull back data from the database. If the code refers to
fields that aren't in the database, this will fail. So the migrations
need to be run before the initial_data script is, to ensure that the
code and data are in sync.