Hardcode Wagtail admin superuser password for simplicity
We weren't seeing any benefit to setting the password via an environment
variable. This change also fixes a problem where the admin password was
not being set when running `initial_data.sh` inside the Docker container
(because `WAGTAIL_ADMIN_PW` had not been added to `.python_env`); now
this is no longer necessary, and the password is set correctly when
running under Docker.