Properly indent regulation appendices paragraphs
We use CSS classes to indent paragraph text. To assign the correct
class, we count the number of dashes in the label to determine
how deeply nested the paragraph is. Appendeices have special
prefixes that are removed before counting the dashes.
e.g. A-1-a becomes a and gets a `level-0` class
e.g. A-2-d-1 becomes d-1 and gets a `level-1` class

https://ghe/eregs/regulations-3000/issues/42#issuecomment-159459