Simplify scripts and docs related to local docker-compose setup
This is certainly subjective, but it seemed like `setup.sh` and
`runserver.sh` were over-doing it by also `source`ing
`mac-virtualbox-init.sh`. For one thing, this could easily lead to
confusion, because the first two scripts `set -e`, whereas the third
does not -- it in fact runs commands _expecting_ non-zero exit codes.
The interaction between these scripts was not ideal.

But moreover it seems best to de-couple the scripts to make for a more
flexible and intentional local Docker-based setup. So this commit
removes the references to `mac-virtualbox-init.sh` in `setup.sh` and
`runserver.sh`, and then updates the documentation to reflect this
change.

And once the `source mac-virtualbox-init.sh` was removed from
`runserver.sh`, there didn't seem to be any good reason to continue
having a `docker` mode in that script. So `runserver.sh` is back to
only being used for standalone mode, and for Docker-based setups
we interact with `docker-compose` directly, which seems more natural.