Convert ImageInset to ContentImage, add 470px width; pre-deprecate Media
This simplifies our image insertion options in FullWidthText to a single
block that can work for both inset (floating) and full-width images.

Adds a new deprecated-blocks.css file for the Wagtail admin to hold
styles dedicated to hiding blocks that are on the path to being
completely removed, like Media; moves the ImageText and LinkBlob hiding
styles there.