workaround to restore Wagtail s3 file overwriting
(See cfgov/v1/storage.py for a very lengthy comment that explains the
need for this commit in great detail.)

Wagtail 1.13 introduced some behavior that broke support for Django file
storage systems that overwrite saved files that have the same name, for
example the django-storages S3 backend we use with the default
AWS_S3_FILE_OVERWRITE=True.

This commit adds an admittedly very hacky workaround to restore
support for this behavior in Wagtail 1.13+. It's not pretty, but it
works.

I've opened a PR to Wagtail to restore this behavior the right way at
https://github.com/wagtail/wagtail/pull/4567; this code in cfgov-refresh
would only be necessary until that is merged in, assuming it is at some
point. The unit tests in this commit are copied from that PR.

This change should fix GHE#2750. @Scotchester and I colllaborated on
this work.

This commit also includes some minor refactoring and simplification of
how S3-related settings are set, to make unit testing easier.