regs3k: Make subpart section_range available in admin
Subparts in the admin tree currently display only subpart title,
isn't helpful if you want to edit a particular section of a large reg.

This elevates the Subpart's `section_range` value from a property to a
model field that gets populated when a subpart is saved.

**Note**

This PR includes a table migration, so you'll need to update your local db:

```
./cfgov/manage.py migrate --noinput
```