Add regulations3k to first-party imports and re-sort imports
We had previously left regulations3k out of the isort list of first-party imports. This will fix that, and will re-sort all of the regulations3k imports appropriately.

The overall change is to place regulations3k imports *with* the rest of cfgov's imports in the regulations3k code, rather than above them.