add PGPASSWORD to .python_env_SAMPLE
In Django < 1.9, manage.py dbshell won't pass the Postgres DB password
automatically when connecting. To avoid getting prompted for the
password, we can set the PGPASSWORD environment variable.

Here we set the sample .python_env file PGPASSWORD to cfpb because that
is the password set in the docker-compose.yml file for the Postgres
container.