Update favicon location and static reference
As part of the Wordpress removal effort, we need to switch to using a
favicon.ico that's part of the Django app. Up till now the favicon.ico
has come from an internal Wordpress theme repo which gets added to the
build artifact via no-drama's `--aux` option.

We also noticed that the existing `<link/>` reference to a favicon in
the site's base template to a static file ('assets/favicon.ico') pointed
to a favicon that, as far as we can tell, did not get placed on the
filesystem as part of the existing Ansible deploy process. That is, the
file did not seem to get placed there via `collectstatic` or some other
build/deploy operation. We don't know how it got there.

To address the confusion we have moved an existing copy of the file
(which is binary equal to the current favicon.ico being served on the
site) so that it will be collected to the top of the STATIC_ROOT.

We'll also need to update the corresponding httpd `Alias` directive so
that '/favicon.ico' on the site will point to this new location.