Ask CFPB: Add count check to avoid index errors
The page's query test (to make sure there are results) was naively
checking the existence of an unevaluated query, instead of checking
that the *evaluated* query produced results.

It's hard to test elasticsearch errors locally, but I've added tests
to check that we handle results of 0 and 1 properly, and, it is hoped,
prevent the index errors.

Since templatetags were recently exposed to testing, this also adds a
test for ask_cfpb's accent stripper.