regs3k: Parse interpretations and insert interp references.
This implements reference parsing and insertion for interpretations,
starting with one of the smaller regs, C.

Included:
* Some refactoring and reorganizing of the many parsing functions.
A grander refactoring will be in order once we can reasonably parse
interpretations for all regs.
* Proper population of the new `subpart_type` field for sorting.
* Bug fixes that were throwing off indentation and boldface.

Testing

This is working well for Reg C, but some of the bigger regs are
still hostile territory. To see how C comes out:

```
./cfgov/manage.py runscript ecfr_importer --script-args 1003
```

This assumes a local cfgov-refresh install and the reg landing page and
a Regulation C reg page set up. You'll need to navigate to the newly
imported C effective-version and uncheck its "draft" box.

Then check this out: <http://localhost:8000/regulations/1003/2/>