Agreements processor: Fix skipped PDFs.
Two bugs were leaving some agreement PDFs missing from the cfgov db.
1. Some folders in the source zip file were empty -- an upstream problem.
2. Some PDFs got skipped because our processor didn't catch PDFs that
had uppercase extension names.

We can't solve the missing files, but we can check for empty folders
and report that.

The root of the uppercase extension problem seems rooted in the fnmatch
utlity, which purports to be case insensitive. In practice, it is not.