Allow breadcrumbs that do not have links
For Regulations3000 we need to allow breadcrumbs that do not provide links in themselves, because we need to include regulation Subparts which do not have urls of their own.

This change adds a check to see if crumb.href or crumb.live exist. the href attribute will indiciate it's a dictionary with an href, the live attribute will indicate that it's a page-like object (get_protected_url uses page.live).

If neither of these exist, we include the breadcrumb as simple text.