enforce case-insensitive tag matching for snippets
The SnippetList organism lets users specify tags that limit which
snippets are returned in the list. The help text for this organism
claims that case-insensitive matching is used, but currently it is not.
This happens to work properly on MySQL but is broken on PostgreSQL.

This change explicitly uses Django's `iexact` matching for
case-insensitive tag matching, and adds a regression unit test to cover
this behavior.